Move the GlobalSubstitution implementation to GlobalSubstitution.cc
